refactor: remove as_information_rules
