+++ cognite/neat/rules/transformers/__init__.py
@@ -1,4 +1,11 @@
 from ._base import RulesPipeline, RulesTransformer
-from ._converters import AssetToInformation, InformationToAsset, InformationToDMS
+from ._converters import AssetToInformation, DMSToInformation, InformationToAsset, InformationToDMS
 
-__all__ = ["RulesTransformer", "RulesPipeline", "InformationToDMS", "InformationToAsset", "AssetToInformation"]
+__all__ = [
+    "RulesTransformer",
+    "RulesPipeline",
+    "InformationToDMS",
+    "InformationToAsset",
+    "AssetToInformation",
+    "DMSToInformation",
+]
